refactor(retrieval): isolate sentence transformer metadata
