+++ tests/unit_tests/_transformers/test_retrieval.py
@@ -251,70 +251,6 @@ def test_direct_standard_export_uses_general_sequence_capabilities(tmp_path):
     assert metadata == {"max_seq_length": tokenizer.model_max_length, "do_lower_case": False}
 
 
-def test_cached_source_model_path_uses_exact_loaded_revision(tmp_path, monkeypatch):
-    from nemo_automodel._transformers import retrieval
-
-    local_root = tmp_path / "local"
-    (local_root / "encoder").mkdir(parents=True)
-    assert retrieval._resolve_cached_source_model_path(
-        str(local_root), SimpleNamespace(), {"subfolder": "encoder"}
-    ) == str(local_root / "encoder")
-    assert retrieval._resolve_cached_source_repository_path(
-        str(local_root), str(local_root / "encoder"), {"subfolder": "encoder"}
-    ) == str(local_root)
-
-    cached_config = tmp_path / "snapshot" / "encoder" / "config.json"
-    cached_config.parent.mkdir(parents=True)
-    cached_config.write_text("{}")
-    cache_lookup = MagicMock(return_value=str(cached_config))
-    monkeypatch.setattr(retrieval, "try_to_load_from_cache", cache_lookup)
-
-    result = retrieval._resolve_cached_source_model_path(
-        "org/model",
-        SimpleNamespace(_commit_hash="exact-commit"),
-        {"cache_dir": "/cache", "revision": "branch", "subfolder": "encoder"},
-    )
-
-    assert result == str(cached_config.parent)
-    assert retrieval._resolve_cached_source_repository_path("org/model", result, {"subfolder": "encoder"}) == str(
-        tmp_path / "snapshot"
-    )
-    cache_lookup.assert_called_once_with(
-        "org/model",
-        "encoder/config.json",
-        cache_dir="/cache",
-        revision="exact-commit",
-    )
-
-
-def test_cache_hub_source_legal_assets_uses_exact_loaded_revision(monkeypatch):
-    from nemo_automodel._transformers import retrieval
-
-    snapshot_download = MagicMock(return_value="/cache/snapshot")
-    monkeypatch.setattr(retrieval, "snapshot_download", snapshot_download)
-
-    result = retrieval._cache_hub_source_legal_assets(
-        "org/model",
-        SimpleNamespace(_commit_hash="exact-commit"),
-        {
-            "cache_dir": "/cache",
-            "revision": "branch",
-            "use_auth_token": "token",
-            "local_files_only": True,
-        },
-    )
-
-    assert result == "/cache/snapshot"
-    snapshot_download.assert_called_once_with(
-        repo_id="org/model",
-        allow_patterns=retrieval._SOURCE_LEGAL_ASSET_PATTERNS,
-        cache_dir="/cache",
-        local_files_only=True,
-        revision="exact-commit",
-        token="token",
-    )
-
-
 def test_direct_standard_export_preserves_cached_source_deployment_limit(tmp_path):
     from nemo_automodel._transformers import retrieval
 
@@ -839,73 +775,6 @@ def test_nemo_bi_encoder_uses_defaults_without_sentence_transformer_metadata():
     config = PretrainedConfig()
 
     assert retrieval._resolve_bi_encoder_options(config, None, None, None) == ("avg", True)
-
-
-def test_sentence_transformer_source_with_unsupported_module_is_rejected(tmp_path):
-    from nemo_automodel._transformers import retrieval
-
-    (tmp_path / "1_Pooling").mkdir()
-    (tmp_path / "modules.json").write_text(
-        json.dumps(
-            [
-                {"idx": 0, "path": "", "type": "sentence_transformers.models.Transformer"},
-                {"idx": 1, "path": "1_Pooling", "type": "sentence_transformers.models.Pooling"},
-                {"idx": 2, "path": "2_Dense", "type": "sentence_transformers.models.Dense"},
-            ]
-        )
-    )
-    (tmp_path / "1_Pooling" / "config.json").write_text(json.dumps({"pooling_mode_mean_tokens": True}))
-
-    with pytest.raises(ValueError, match="exact supported module stack"):
-        retrieval._load_sentence_transformer_wrapper_options(str(tmp_path), {})
-
-
-def test_sentence_transformer_source_excluding_prompts_is_rejected(tmp_path):
-    from nemo_automodel._transformers import retrieval
-
-    (tmp_path / "1_Pooling").mkdir()
-    (tmp_path / "modules.json").write_text(
-        json.dumps(
-            [
-                {"idx": 0, "path": "", "type": "sentence_transformers.models.Transformer"},
-                {"idx": 1, "path": "1_Pooling", "type": "sentence_transformers.models.Pooling"},
-            ]
-        )
-    )
-    (tmp_path / "1_Pooling" / "config.json").write_text(
-        json.dumps({"pooling_mode_mean_tokens": True, "include_prompt": False})
-    )
-
-    with pytest.raises(ValueError, match="include_prompt=False"):
-        retrieval._load_sentence_transformer_wrapper_options(str(tmp_path), {})
-
-
-def test_sentence_transformer_source_lowercasing_is_rejected(tmp_path):
-    from nemo_automodel._transformers import retrieval
-
-    (tmp_path / "1_Pooling").mkdir()
-    (tmp_path / "modules.json").write_text(
-        json.dumps(
-            [
-                {"idx": 0, "path": "", "type": "sentence_transformers.models.Transformer"},
-                {"idx": 1, "path": "1_Pooling", "type": "sentence_transformers.models.Pooling"},
-            ]
-        )
-    )
-    (tmp_path / "1_Pooling" / "config.json").write_text(json.dumps({"pooling_mode_mean_tokens": True}))
-    (tmp_path / "sentence_bert_config.json").write_text(json.dumps({"do_lower_case": True}))
-
-    with pytest.raises(ValueError, match="do_lower_case=True"):
-        retrieval._load_sentence_transformer_wrapper_options(str(tmp_path), {})
-
-
-def test_sentence_transformer_metadata_load_fails_closed_on_unexpected_hub_error(monkeypatch):
-    from nemo_automodel._transformers import retrieval
-
-    monkeypatch.setattr(retrieval, "hf_hub_download", MagicMock(side_effect=OSError("offline")))
-
-    with pytest.raises(RuntimeError, match="Unable to load Sentence Transformers metadata modules.json"):
-        retrieval._load_sentence_transformer_json("org/model", "modules.json", {})
 
 
 def test_extract_submodel_ministral_embedding_from_local_vlm_converts_to_supported_backbone(tmp_path):
